TG/coruja/backend/requirements.txt
Flask==2.3.3
psycopg2-binary==2.9.9
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-JWT-Extended==4.5.3
Flask-Bcrypt==1.0.1
python-dotenv==1.0.0
Flask-CORS==4.0.0
PyJWT==2.8.0
httpx==0.25.2
fastapi==0.103.1
uvicorn==0.23.2
python-multipart==0.0.6
ollama==0.1.6    